+++ permission_handler/ios/Classes/strategies/AudioVideoPermissionStrategy.m
@@ -13,17 +13,13 @@
     if (permission == PermissionGroupCamera) {
         #if PERMISSION_CAMERA
         return [AudioVideoPermissionStrategy permissionStatus:AVMediaTypeVideo];
-        #else
-        return PermissionStatusUnknown;
         #endif
     } else if (permission == PermissionGroupMicrophone) {
         #if PERMISSION_MICROPHONE
         return [AudioVideoPermissionStrategy permissionStatus:AVMediaTypeAudio];
-        #else
-        return PermissionStatusUnknown;
         #endif
     }
-    return PermissionStatusUnknown;
+    return PermissionStatusNotDetermined;
 }
 
 - (ServiceStatus)checkServiceStatus:(PermissionGroup)permission {
@@ -33,7 +29,7 @@
 - (void)requestPermission:(PermissionGroup)permission completionHandler:(PermissionStatusHandler)completionHandler {
     PermissionStatus status = [self checkPermissionStatus:permission];
 
-    if (status != PermissionStatusUnknown) {
+    if (status != PermissionStatusNotDetermined) {
         completionHandler(status);
         return;
     }
@@ -44,18 +40,18 @@
         #if PERMISSION_CAMERA
         mediaType = AVMediaTypeVideo;
         #else
-        completionHandler(PermissionStatusUnknown);
+        completionHandler(PermissionStatusNotDetermined);
         return;
         #endif
     } else if (permission == PermissionGroupMicrophone) {
         #if PERMISSION_MICROPHONE
         mediaType = AVMediaTypeAudio;
         #else
-        completionHandler(PermissionStatusUnknown);
+        completionHandler(PermissionStatusNotDetermined);
         return;
         #endif
     } else {
-        completionHandler(PermissionStatusUnknown);
+        completionHandler(PermissionStatusNotDetermined);
         return;
     }
 
@@ -73,7 +69,7 @@
 
     switch (status) {
         case AVAuthorizationStatusNotDetermined:
-            return PermissionStatusUnknown;
+            return PermissionStatusNotDetermined;
         case AVAuthorizationStatusRestricted:
             return PermissionStatusRestricted;
         case AVAuthorizationStatusDenied:
@@ -82,7 +78,7 @@
             return PermissionStatusGranted;
     }
 
-    return PermissionStatusUnknown;
+    return PermissionStatusNotDetermined;
 }
 
 @end