+++ permission_handler_platform_interface/lib/src/permission_handler_platform_interface.dart
@@ -1,6 +1,6 @@
 part of permission_handler_platform_interface;
 
-/// The interface that implementations of permission_handler must implement.
+/// The interface that implementations of `permission_handler` must implement.
 ///
 /// Platform implementations should extend this class rather than implement it
 /// as `permission_handler` does not consider newly added methods to be
@@ -21,71 +21,67 @@ abstract class PermissionHandlerPlatform extends PlatformInterface {
   /// Defaults to [MethodChannelPermissionHandler].
   static PermissionHandlerPlatform get instance => _instance;
 
-  /// Platform-specific plugins should set this with their own platform-specific
-  /// class that extends [PermissionHandlerPlatform] when they register themselves.
+  /// Platform-specific plugins should set this with their own
+  /// platform-specific class that extends
+  /// [PermissionHandlerPlatform] when they register themselves.
   static set instance(PermissionHandlerPlatform instance) {
     PlatformInterface.verifyToken(instance, _token);
     _instance = instance;
   }
 
-  /// Check current permission status.
-  ///
-  /// Returns a [Future] containing the current permission status for the
-  /// supplied [PermissionGroup].
-  Future<PermissionStatus> checkPermissionStatus(PermissionGroup permission) {
+  /// Checks the current status of the given [Permission].
+  Future<PermissionStatus> checkPermissionStatus(Permission permission) {
     throw UnimplementedError(
         'checkPermissionStatus() has not been implemented.');
   }
 
-  /// Check current service status.
+  /// Checks the current status of the service associated with the given
+  /// [Permission].
   ///
-  /// Returns a [Future] containing the current service status for the supplied
-  /// [PermissionGroup].
-  ///
-  /// Notes about specific PermissionGroups:
-  /// - **PermissionGroup.phone**
+  /// Notes about specific permissions:
+  /// - **[Permission.phone]**
   ///   - Android:
   ///     - The method will return [ServiceStatus.notApplicable] when:
-  ///       1. the device lacks the TELEPHONY feature
-  ///       1. TelephonyManager.getPhoneType() returns PHONE_TYPE_NONE
-  ///       1. when no Intents can be resolved to handle the `tel:` scheme
+  ///       - the device lacks the TELEPHONY feature
+  ///       - TelephonyManager.getPhoneType() returns PHONE_TYPE_NONE
+  ///       - when no Intents can be resolved to handle the `tel:` scheme
   ///     - The method will return [ServiceStatus.disabled] when:
-  ///       1. the SIM card is missing
+  ///       - the SIM card is missing
   ///   - iOS:
   ///     - The method will return [ServiceStatus.notApplicable] when:
-  ///       1. the native code can not find a handler for the `tel:` scheme
+  ///       - the native code can not find a handler for the `tel:` scheme
   ///     - The method will return [ServiceStatus.disabled] when:
-  ///       1. the mobile network code (MNC) is either 0 or 65535. See
+  ///       - the mobile network code (MNC) is either 0 or 65535. See
   ///          https://stackoverflow.com/a/11595365 for details
   ///   - **PLEASE NOTE that this is still not a perfect indication** of the
-  ///     devices' capability to place & connect phone calls
-  ///     as it also depends on the network condition.
-  Future<ServiceStatus> checkServiceStatus(PermissionGroup permission) {
+  ///     device's capability to place & connect phone calls as it also depends
+  ///     on the network condition.
+  Future<ServiceStatus> checkServiceStatus(Permission permission) {
     throw UnimplementedError('checkServiceStatus() has not been implemented.');
   }
 
-  /// Open the App settings page.
+  /// Opens the app settings page.
   ///
-  /// Returns [true] if the app settings page could be opened,
-  /// otherwise [false] is returned.
+  /// Returns [true] if the app settings page could be opened, otherwise
+  /// [false].
   Future<bool> openAppSettings() {
     throw UnimplementedError('openAppSettings() has not been implemented.');
   }
 
-  /// Request the user for access to the supplied list of permissiongroups.
+  /// Requests the user for access to the supplied list of [Permission]s, if
+  /// they have not already been granted before.
   ///
-  /// Returns a [Map] containing the status per requested permissiongroup.
-  Future<Map<PermissionGroup, PermissionStatus>> requestPermissions(
-      List<PermissionGroup> permissions) {
+  /// Returns a [Map] containing the status per requested [Permission].
+  Future<Map<Permission, PermissionStatus>> requestPermissions(
+      List<Permission> permissions) {
     throw UnimplementedError('requestPermissions() has not been implemented.');
   }
 
-  /// Request to see if you should show a rationale for requesting permission.
+  /// Checks if you should show a rationale for requesting permission.
   ///
   /// This method is only implemented on Android, calling this on iOS always
   /// returns [false].
-  Future<bool> shouldShowRequestPermissionRationale(
-      PermissionGroup permission) {
+  Future<bool> shouldShowRequestPermissionRationale(Permission permission) {
     throw UnimplementedError(
         'shouldShowRequestPermissionRationale() has not been implemented.');
   }